backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# OpenCASCADE Python bindings
pythonocc-core>=7.7.0

# Numerical computing
numpy>=1.24.0

# glTF export
pygltflib>=1.16.0

# Claude API
anthropic>=0.8.0

# Data validation and settings
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Additional utilities
python-dotenv>=1.0.0
aiofiles>=23.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Code quality
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0
